+++ sdk/storage/storage-blob/src/index.ts
@@ -6,21 +6,21 @@ import { RestError } from "@azure/core-http";
 export { PollerLike, PollOperationState } from "@azure/core-lro";
 export * from "./BlobServiceClient";
 export * from "./Clients";
-export * from "./AccountSASPermissions";
-export * from "./AccountSASResourceTypes";
-export * from "./AccountSASServices";
-export * from "./AccountSASSignatureValues";
+export * from "./sas/AccountSASPermissions";
+export * from "./sas/AccountSASResourceTypes";
+export * from "./sas/AccountSASServices";
+export * from "./sas/AccountSASSignatureValues";
 export * from "./BlobBatch";
 export * from "./BlobBatchClient";
 export * from "./BatchResponse";
-export * from "./BlobSASPermissions";
-export * from "./BlobSASSignatureValues";
+export * from "./sas/BlobSASPermissions";
+export * from "./sas/BlobSASSignatureValues";
 export * from "./StorageBrowserPolicyFactory";
-export * from "./ContainerSASPermissions";
+export * from "./sas/ContainerSASPermissions";
 export * from "./credentials/AnonymousCredential";
 export * from "./credentials/Credential";
 export * from "./credentials/StorageSharedKeyCredential";
-export { SasIPRange } from "./SasIPRange";
+export { SasIPRange } from "./sas/SasIPRange";
 export { Range } from "./Range";
 export {
   BlockBlobTier,
@@ -38,7 +38,7 @@ export * from "./policies/AnonymousCredentialPolicy";
 export * from "./policies/CredentialPolicy";
 export * from "./StorageRetryPolicyFactory";
 export * from "./policies/StorageSharedKeyCredentialPolicy";
-export * from "./SASQueryParameters";
+export * from "./sas/SASQueryParameters";
 export { CommonOptions } from "./StorageClient";
 export * from "./generatedModels";
 export {
