+++ sdk/storage/storage-blob/src/sas/AccountSASPermissions.ts
