+++ sdk/storage/storage-blob/src/sas/AccountSASResourceTypes.ts
